h1.8

+++ src/main/java/com/chen/sbbus/entity/Station.java
@@ -17,10 +17,10 @@ public class Station {
 //    @TableId(value="id",type= IdType.AUTO)//自增id
     Integer id;
     private String name;
-    private Double longitude;//经度最高
+    private float longitude;//经度最高
     @TableField("E_W")
     private String E_W;//经度最低
-    private Double latitude;  //纬度最高
+    private float latitude;  //纬度最高
     @TableField("S_N")
     private String S_N;//纬度最低
